Date: Fri, 20 Aug 2004 08:48:38 +0400 From: Daniel.Lostotter@vcmain.hq.c5.army.mil To: webmaster@hebcal.com Subject: candle lighting time for Baghdad
Sir,
We are Jewish soldiers stationed in Baghdad, Iraq and were wondering if you
would consider adding Baghdad to your list of cities for candle lighting
times.  It is a wondrous event to be Jews in Baghdad observing our faith and
to be recognized on your website would be an honor.

Thank you on behalf of the Victory Base Jewish Community.

B'Shalom

Daniel Lostotter