Fix the `numSuffix` calculation for numbers with more than 2 digits.
For numbers in the "teens" e.g. 13, this function would take a shortcut to return "th", but it did not correctly apply that same shortcut for numbers with more than 2 digits, e.g. 113, which would incorrectly return "rd".

Although the Omer never reaches triple digits, who knows where this function might be used in the future. Might as well fix it now.

Also fix the header guard to match the `#ifndef` check.