Basic Zemanim support.
Changed Candlelighting to say Light Candles Before or Light Candles
After. Added candle lighting on the second night of Yom Tov, fixing
issue #94.

When -Z is included on the command line, a basic set of
Zemanim is included for each day, namely:

(ZMAN_SUNRISE | ZMAN_SZKS | ZMAN_TEFILAH |
 ZMAN_CHATZOT |
 ZMAN_MINCHA_GEDOLA | ZMAN_MINCHA_KETANA |
 ZMAN_PLAG_HAMINCHA | ZMAN_SUNSET | ZMAN_TZAIT_42)

iso8859-8 text is not included for these but should be added. (Or maybe
it's time to switch to UTF-8?)