German translation po file added
Translations into German which follow the common conventions for transliteration in liturgical context:

The softly spoken "ב" and the "ו" spoken as a consonant both transliterate to "w", e.g. "Wajikra".
The "י" transliterates to "j" if pronounced as a consonant, while "i" when pronounced as a vocal, e.g.: "Wajikra".
The "ב" at the end of a word is transliterated "w", e.g. "Jom Tow".
Except for "Jom haSikaron", the modern rememberance days are translated to "... - Tag", while "Aliyah" is not translated, as no one in the community calls "Aliyah" other than "Aliyah" (it is also not called "immigration day", "elevation day" or "return day" or anything like that in English).